refactor: add alias of common in components
