+++ frontend/src/components/Common/Overlay/index.jsx
@@ -1,1 +1,1 @@
-export {default} from './Overlay';
+export { default } from './Overlay';
