--- frontend/src/components/Common/Slider/Slider/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './Slider';
