--- frontend/src/components/Common/Slider/SliderItem/index.jsx
@@ -1,1 +1,0 @@
-export { default } from 'SliderItem';
