+++ frontend/src/components/Common/Slider/index.jsx
@@ -1,4 +1,1 @@
-import Slider from './Slider';
-import SliderItem from './SliderItem';
-
-export { Slider, SliderItem };
+export { default } from './Slider';
