+++ frontend/src/components/Common/SliderItem/SliderItem.jsx
