+++ frontend/src/components/Common/SliderItem/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SliderItem';
