+++ frontend/src/components/Common/index.jsx
@@ -1,0 +1,29 @@
+import Box from './Box';
+import Avatar from './Avatar';
+import Button from './Button';
+import Card from './Card';
+import Chip from './Chip';
+import Divider from './Divider';
+import IconButton from './IconButton';
+import Modal from './Modal';
+import PageStatus from './PageStatus';
+import Slider from './Slider';
+import SliderItem from './SliderItem';
+import TextField from './TextField';
+import Typography from './Typography';
+
+export {
+    Box,
+    Avatar,
+    Button,
+    Card,
+    Chip,
+    Divider,
+    IconButton,
+    Modal,
+    PageStatus,
+    Slider,
+    SliderItem,
+    TextField,
+    Typography,
+};
